ENH: Part 1 - Move current IO readers and writers to legacy
* Moves all API headers
* Moves readers and writers classes that use `gdf_*`
* Moves all IO readers and writers gtests
