+++ cpp/include/cudf/legacy/io_functions.hpp
@@ -18,7 +18,8 @@
 
 #include <memory>
 
-#include "cudf.h"
+#include <cudf/cudf.h>
+#include <cudf/types.hpp>
 #include <cudf/legacy/table.hpp>
 
 namespace cudf {
