+++ cpp/include/cudf/legacy/io_readers.hpp
@@ -21,7 +21,8 @@
 #include <utility>
 #include <vector>
 
-#include "cudf.h"
+#include <cudf/cudf.h>
+#include <cudf/types.hpp>
 #include <cudf/legacy/table.hpp>
 
 // Forward declarations
