+++ cpp/include/cudf/legacy/io_types.hpp
@@ -21,8 +21,10 @@
 #include <memory>
 #include <utility>
 
+#include <cudf/cudf.h>
+#include <cudf/types.hpp>
 #include <cudf/legacy/table.hpp>
-#include "cudf.h"
+#include <cudf/legacy/io_types.h>
 
 // Forward declarations
 namespace arrow { namespace io {  class RandomAccessFile; } }
