+++ cpp/src/io/csv/legacy/csv_reader_impl.cu
@@ -41,8 +41,8 @@
 #include <nvstrings/NVStrings.h>
 
 #include <io/comp/io_uncomp.h>
-#include <io/cuio_common.hpp>
-#include <io/utilities/parsing_utils.cuh>
+#include <io/legacy/cuio_common.hpp>
+#include <io/utilities/legacy/parsing_utils.cuh>
 
 using std::vector;
 using std::string;
