+++ cpp/src/io/legacy/cuio_common.cpp
