+++ cpp/src/io/legacy/io_functions.cpp
