+++ cpp/src/io/orc/legacy/orc_reader_impl.hpp
