+++ cpp/src/io/orc/legacy/orc_writer_impl.hpp
