+++ cpp/tests/io/legacy/avro_test.cpp
@@ -20,7 +20,7 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include <tests/io/io_test_utils.hpp>
+#include <tests/io/legacy/io_test_utils.hpp>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 TempDirTestEnvironment *const temp_env = static_cast<TempDirTestEnvironment *>(
