+++ cpp/tests/io/legacy/csv_test.cu
@@ -26,7 +26,7 @@
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include <tests/io/io_test_utils.hpp>
+#include <tests/io/legacy/io_test_utils.hpp>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 #include <arrow/io/api.h>
