+++ cpp/tests/io/legacy/io_test_utils.cpp
